+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add `javaLib` configuration closure:
     - disableGradleMetadata() - disables gradle metadata publishing
     - bonArtifactId and bomDescription - corrects bom pom when platform declared in the root project
+    - disableJavadocPublish() and disableSourcesPublish() - disable javadoc and sources publish, enabled by default
     
 
 ### 2.1.0 (2020-01-19)